+++ sdk/core/azure-core/CHANGELOG.md
@@ -4,7 +4,14 @@
 
 ### Features Added
 
+- Added `SyncPoller.getFinalResult(Duration)` to get the final result of a long-running operation with a timeout
+  period. Allows for a single call rather than calling both `SyncPoller.waitForCompletion(Duration)` and
+  `SyncPoller.getFinalResult()`.
+
 ### Breaking Changes
+
+- `SyncPoller.waitForCompletion(Duration)` and `SyncPoller.waitUntil(Duration, LongRunningOperationStatus)` now
+  throw an exception if the polling operation doesn't complete or reach the status within the give duration.
 
 ### Bugs Fixed